refactor : SwitchButton -> IconButton 으로 네이밍 변경
iconSrc를 그대로 렌더링 하는 버튼으로 변경
