+++ frontend/src/components/@common/IconButton/IconButton.stories.tsx
@@ -1,0 +1,21 @@
+import type { Meta, StoryObj } from '@storybook/react-webpack5';
+import IconButton from './IconButton';
+
+const meta: Meta<typeof IconButton> = {
+  title: 'Common/IconButton',
+  component: IconButton,
+  parameters: {
+    layout: 'centered',
+  },
+  tags: ['autodocs'],
+};
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const Default: Story = {
+  args: {
+    iconSrc: '/images/statistics-icon.svg',
+    onClick: () => console.log('IconButton clicked'),
+  },
+};
